addon classname change
addon classname change & minor work on impact hearing
