--- addons/orbis_rhs_tank_config/XEH_PREP.sqf
@@ -1,1 +1,0 @@
-orbis_fnc_tankImpactHearing = compile preprocessFileLineNumbers 'orbis_rhs_tank_config\scripts\fnc_tankImpactHearing.sqf';
