--- addons/orbis_rhs_tank_config/config.cpp
@@ -1,4 +1,0 @@
-#include "CfgPatches.hpp"
-// #include "CfgEventHandlers.hpp"
-
-#include "CfgVehicles.hpp"
