+++ addons/orbis_tank_config/CfgPatches.hpp
@@ -1,6 +1,6 @@
 class CfgPatches {
-	class orbis_rhs_tank_config {
-		name = "Orbis RHS Tank Config";
+	class orbis_tank_config {
+		name = "Orbis Tank Config";
 		author = "Orbis2358";
 		requiredVersion = 0.1;
 		requiredAddons[] = {"RHS_US_A2Port_Armor", "rhsusf_c_m1a2", "rhsusf_c_m113"};
