+++ addons/orbis_tank_config/XEH_PREP.sqf
@@ -1,0 +1,1 @@
+orbis_tank_fnc_tankImpactHearing = compile preprocessFileLineNumbers 'orbis_tank_config\scripts\fnc_tankImpactHearing.sqf';
