+++ addons/orbis_tank_config/config.cpp
@@ -1,0 +1,4 @@
+#include "CfgPatches.hpp"
+#include "CfgEventHandlers.hpp"
+
+#include "CfgVehicles.hpp"
